dev-requirements.txt
python-dotenv
google-search-results
serpapi
